cleaner stresstest, process should be niced somehow
