--- src/strategies/LocalFileSystem/test/compileAndRunMockController.bat
@@ -1,3 +1,0 @@
-cls
-g++ -std=c++11 -ID:\Projects\PJON\src  MockStepperController.cpp -o MockStepperController.cpp -o MockStepperController
-MockStepperController %1 %2
