--- src/strategies/LocalFileSystem/test/compileAndRunMockController.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-printf "\033c"
-g++ -std=c++11 -I/Users/det/Documents/Arduino/Libraries/PJON/src  MockStepperController.cpp -o MockStepperController
-./MockStepperController $1
