--- src/strategies/LocalFileSystem/test/compileAndRunMockSender.bat
@@ -1,3 +1,0 @@
-cls
-g++ -std=c++11 -ID:\Projects\PJON\src  MockSender.cpp -o MockSender
-MockSender
