+++ src/strategies/LocalFileSystem/test/compileAndRunTestClient.bat
@@ -1,0 +1,3 @@
+cls
+g++ -std=c++11 -ID:\Projects\PJON\src  TestClient.cpp -o TestClient
+TestClient %1 %2
