+++ src/strategies/LocalFileSystem/test/compileAndRunTestSender.bat
@@ -1,0 +1,3 @@
+cls
+g++ -std=c++11 -ID:\Projects\PJON\src  TestSender.cpp -o TestSender
+TestSender %1
